Поправил контроллеры, классы сущностей и сервисы. Убрал многие неиспользуемые импорты, подравнял код, убрал некоторые неиспользуемые методы. Добавил отображение ошибок входа и информации о выходе на login-страницу. Реализовал работу с атрибутами сессии для бана. Сделал валидацию бана при его сохранении. Исправил некорректные функции в utility-классах

+++ src/main/java/org/forum/main/services/implementations/DislikeServiceImpl.java
@@ -36,6 +36,9 @@ public class DislikeServiceImpl implements DislikeService {
 
     @Override
     public void save(Message dislikedMessage, User userWhoDisliked) {
+        if (dislikedMessage.getUserWhoPosted().getId().equals(userWhoDisliked.getId())) {
+            return;
+        }
         if (dislikedMessage.containsLikedUserById(userWhoDisliked.getId())) {
             likeRepository.delete(likeRepository.findByMessageIdAndUserId(dislikedMessage.getId(), userWhoDisliked.getId())
                     .orElseThrow(
